Tags

+++ db/migrate/20190410074423_add_latitude_to_shops.rb
@@ -1,5 +1,5 @@
 class AddLatitudeToShops < ActiveRecord::Migration[5.2]
   def change
-    add_column :shops, :latitude, :integer
+    add_column :shops, :latitude, :float
   end
 end
