+++ db/migrate/20190410074444_add_longitude_to_shops.rb
@@ -1,5 +1,5 @@
 class AddLongitudeToShops < ActiveRecord::Migration[5.2]
   def change
-    add_column :shops, :longitude, :integer
+    add_column :shops, :longitude, :float
   end
 end
